% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Extraer_AGEB.R
\name{extraer_ageb}
\alias{extraer_ageb}
\title{Title}
\usage{
extraer_ageb(data_ageb, data_loc, localidad)
}
\arguments{
\item{data_ageb}{es la base de ageb del estado}

\item{data_loc}{es la base de localidades del estado}

\item{localidad}{es la localidad del interes}
}
\value{
un objeto sf
}
\description{
Title
}
\author{
Casandra Ailicec Robles Cruz
}
